+++ src/providers/job_definition_provider.ts
@@ -1,11 +1,11 @@
 import * as vscode from "vscode";
 import { JobParser } from "../job_parsing/job_parser";
-import { JobDefinitionManager } from "../job_parsing/job_definition_manager";
+import { JobManager } from "../job_parsing/job_manager";
 import { ProjectTemplateParser } from "../project_template_parsing/project_template_parser";
 import path = require("path");
 
 export class JobDefinitionProvider implements vscode.DefinitionProvider {
-	constructor(private readonly job_manager: JobDefinitionManager) {}
+	constructor(private readonly job_manager: JobManager) {}
 
 	provideDefinition(
 		document: vscode.TextDocument,
@@ -20,9 +20,6 @@ export class JobDefinitionProvider implements vscode.DefinitionProvider {
 				let parent_job = this.job_manager.get_job_with_name(parent_name);
 				if (parent_job) {
 					let attribute = parent_job.get_job_name_attribute();
-					if (!attribute) {
-						throw new Error("Parent attribute doesn't exist>?!?!?!?!");
-					}
 					return new vscode.Location(attribute.location.document, attribute.location.range);
 				}
 			}
@@ -33,9 +30,6 @@ export class JobDefinitionProvider implements vscode.DefinitionProvider {
 				let job = this.job_manager.get_job_with_name(job_name);
 				if (job) {
 					let attribute = job.get_job_name_attribute();
-					if (!attribute) {
-						throw new Error("Parent attribute doesn't exist>?!?!?!?!");
-					}
 					return new vscode.Location(attribute.location.document, attribute.location.range);
 				}
 			}