+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 All notable changes to the "zuulplugin" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.2.1]
+
+- Job attributes and project-template attributes now have accurate locations in vscode space.
+- Added the ability to rename job-names in a single click.
 
 ## [0.2.0]